refactor: move merkle root calculation to MerkleRoot
