+++ core/src/main/java/org/bitcoinj/quorums/SimplifiedQuorumList.java
@@ -299,7 +299,7 @@ public class SimplifiedQuorumList extends Message {
 
             if (!cbtx.getMerkleRootQuorums().isZero() &&
                     !commitmentHashes.isEmpty() &&
-                    !cbtx.getMerkleRootQuorums().equals(calculateMerkleRoot(commitmentHashes)))
+                    !cbtx.getMerkleRootQuorums().equals(MerkleRoot.calculateMerkleRoot(commitmentHashes)))
                 throw new MasternodeListDiffException("MerkleRoot of quorum list does not match coinbaseTx - " + commitmentHashes.size(), true, false, false, true);
 
             return true;
@@ -323,12 +323,9 @@ public class SimplifiedQuorumList extends Message {
             }
         });
 
-        if (!merkleRootQuorums.isZero() &&
-                !commitmentHashes.isEmpty() &&
-                !merkleRootQuorums.equals(calculateMerkleRoot(commitmentHashes)))
-            return false;
-
-        return true;
+        return merkleRootQuorums.isZero() ||
+                commitmentHashes.isEmpty() ||
+                merkleRootQuorums.equals(MerkleRoot.calculateMerkleRoot(commitmentHashes));
     }
 
     public Sha256Hash calculateMerkleRoot() {
@@ -347,69 +344,10 @@ public class SimplifiedQuorumList extends Message {
                 }
             });
 
-            return calculateMerkleRoot(commitmentHashes);
+            return MerkleRoot.calculateMerkleRoot(commitmentHashes);
         } finally {
             lock.unlock();
         }
-    }
-
-    public static Sha256Hash calculateMerkleRoot(List<Sha256Hash> hashes) {
-        List<byte[]> tree = buildMerkleTree(hashes);
-        return Sha256Hash.wrap(tree.get(tree.size() - 1));
-    }
-
-    private static List<byte[]> buildMerkleTree(List<Sha256Hash> hashes) {
-        // The Merkle root is based on a tree of hashes calculated from the masternode list proRegHash:
-        //
-        //     root
-        //      / \
-        //   A      B
-        //  / \    / \
-        // t1 t2 t3 t4
-        //
-        // The tree is represented as a list: t1,t2,t3,t4,A,B,root where each
-        // entry is a hash.
-        //
-        // The hashing algorithm is double SHA-256. The leaves are a hash of the serialized contents of the transaction.
-        // The interior nodes are hashes of the concenation of the two child hashes.
-        //
-        // This structure allows the creation of proof that a transaction was included into a block without having to
-        // provide the full block contents. Instead, you can provide only a Merkle branch. For example to prove tx2 was
-        // in a block you can just provide tx2, the hash(tx1) and B. Now the other party has everything they need to
-        // derive the root, which can be checked against the block header. These proofs aren't used right now but
-        // will be helpful later when we want to download partial block contents.
-        //
-        // Note that if the number of transactions is not even the last tx is repeated to make it so (see
-        // tx3 above). A tree with 5 transactions would look like this:
-        //
-        //         root
-        //        /     \
-        //       1        5
-        //     /   \     / \
-        //    2     3    4  4
-        //  / \   / \   / \
-        // t1 t2 t3 t4 t5 t5
-        ArrayList<byte[]> tree = new ArrayList<byte[]>();
-        // Start by adding all the hashes of the transactions as leaves of the tree.
-        for (Sha256Hash hash : hashes) {
-            tree.add(hash.getBytes());
-        }
-        int levelOffset = 0; // Offset in the list where the currently processed level starts.
-        // Step through each level, stopping when we reach the root (levelSize == 1).
-        for (int levelSize = hashes.size(); levelSize > 1; levelSize = (levelSize + 1) / 2) {
-            // For each pair of nodes on that level:
-            for (int left = 0; left < levelSize; left += 2) {
-                // The right hand node can be the same as the left hand, in the case where we don't have enough
-                // transactions.
-                int right = Math.min(left + 1, levelSize - 1);
-                byte[] leftBytes = Utils.reverseBytes(tree.get(levelOffset + left));
-                byte[] rightBytes = Utils.reverseBytes(tree.get(levelOffset + right));
-                tree.add(Utils.reverseBytes(hashTwice(leftBytes, 0, 32, rightBytes, 0, 32)));
-            }
-            // Move to the next level.
-            levelOffset += levelSize;
-        }
-        return tree;
     }
 
     public void addQuorum(Quorum quorum) {
